Moved check/contour_count from `GFonts` profile
to `Universal` profile.
(issue #3491)
